+++ rotkehlchen/chain/ethereum/modules/paladin/constants.py
@@ -1,0 +1,8 @@
+from typing import Final
+
+from rotkehlchen.chain.evm.types import string_to_evm_address
+
+CPT_PALADIN: Final = 'paladin'
+
+PALADIN_MERKLE_DISTRIBUTOR_V2: Final = string_to_evm_address('0xce6dc32252d85e2e955Bfd3b85660917F040a933')  # noqa: E501
+CLAIMED: Final = b'\x9aSv\xf7\xdc\xf8c\x1c+bI\xc9\xbe\xc3\xd7\x15\xcb\x97\xbd\xd4\xc8-\x92\xe5]\x14\x7fkN\xeaA\x97'  # noqa: E501